requirements.txt
streamlit>=1.45.0
requests>=2.32.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn>=0.32.0
